# Backend/internal/interfaces/api/favorite_handler.go
package api

import (
	"github.com/gin-gonic/gin"

	"uniStore/Backend/internal/domain/services"
)

// FavoriteHandler handles HTTP requests related to favorites
type FavoriteHandler struct {
	favoriteService services.FavoriteService
}

// NewFavoriteHandler creates a new favorite handler
func NewFavoriteHandler(favoriteService services.FavoriteService) *FavoriteHandler {
	return &FavoriteHandler{
		favoriteService: favoriteService,
	}
}

// GetFavorite retrieves a user's favorites
// @Summary Get user's favorites
// @Description Returns the list of games favorited by the user
// @Tags Favorites
// @Accept json
// @Produce json
// @Param user_id path int true "User ID"
// @Success 200 {array} dto.GameDTO "List of favorite games"
// @Failure 400 {object} map[string]interface{} "Invalid user ID"
// @Failure 401 {object} map[string]interface{} "Unauthorized"
// @Failure 403 {object} map[string]interface{} "Forbidden"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Security ApiKeyAuth
// @Router /api/v1/favorite/{user_id} [get]
func (h *FavoriteHandler) GetFavorite(c *gin.Context) {
	// Implementation to be added
}

// AddGameToFavorite adds a game to a user's favorites
// @Summary Add game to favorites
// @Description Adds a game to a user's favorites list
// @Tags Favorites
// @Accept json
// @Produce json
// @Param user_id path int true "User ID"
// @Param game_id path int true "Game ID"
// @Success 200 {object} map[string]interface{} "Game added to favorites successfully"
// @Failure 400 {object} map[string]interface{} "Invalid input"
// @Failure 401 {object} map[string]interface{} "Unauthorized"
// @Failure 403 {object} map[string]interface{} "Forbidden"
// @Failure 404 {object} map[string]interface{} "Game not found"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Security ApiKeyAuth
// @Router /api/v1/favorite/{user_id}/add/{game_id} [post]
func (h *FavoriteHandler) AddGameToFavorite(c *gin.Context) {
	// Implementation to be added
}

// RemoveGameFromFavorite removes a game from a user's favorites
// @Summary Remove game from favorites
// @Description Removes a game from a user's favorites list
// @Tags Favorites
// @Accept json
// @Produce json
// @Param user_id path int true "User ID"
// @Param game_id path int true "Game ID"
// @Success 200 {object} map[string]interface{} "Game removed from favorites successfully"
// @Failure 400 {object} map[string]interface{} "Invalid input"
// @Failure 401 {object} map[string]interface{} "Unauthorized"
// @Failure 403 {object} map[string]interface{} "Forbidden"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Security ApiKeyAuth
// @Router /api/v1/favorite/{user_id}/remove/{game_id} [delete]
func (h *FavoriteHandler) RemoveGameFromFavorite(c *gin.Context) {
	// Implementation to be added
}

// ClearFavorite clears a user's favorites
// @Summary Clear all favorites
// @Description Removes all games from a user's favorites list
// @Tags Favorites
// @Accept json
// @Produce json
// @Param user_id path int true "User ID"
// @Success 200 {object} map[string]interface{} "Favorites cleared successfully"
// @Failure 400 {object} map[string]interface{} "Invalid user ID"
// @Failure 401 {object} map[string]interface{} "Unauthorized"
// @Failure 403 {object} map[string]interface{} "Forbidden"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Security ApiKeyAuth
// @Router /api/v1/favorite/{user_id}/clear [delete]
func (h *FavoriteHandler) ClearFavorite(c *gin.Context) {
	// Implementation to be added
}